Python/requirements.txt
picamera
pyzmq
